Add CreditManager
CreditManager is the model in the system that knows how to transfer
funds around and manage our various accounts.  Added unit tests for
credit manager, and enabled the high level feature that ensures that
money is moved around the system (give money to a school).  We can now
continue down the path of integration testing the money transfer system.