Admin New Orders
created controller action to create a new order, set the products and
quantities and line_items. then move the order along to the completed
stage, set the shipping method to ship_to_school, and redirect to the
admin orders page. created a migration to add school_id to spree_orders
so that we know what school the order is for