check subdomain against current_school instead person.school
A student can belong to multiple schools, so the person#school
method just gets the first school. The problem with this is that
the school that is returned, might not be the current_school, or
school that they user actually attends. So, they are unable to
purchase from the store.