Scope searched OtuCode by the current_person
The issue before was that since codes were not
unique, when we searched all OtuCodes for a
specific code, we had a good chance of pulling
back a code that was for a different user.

This method isn't bulletproof. There is still a
possibility for the same user to have multiple
otu_codes that have the same code, but it's highly
unlikely, and the underlying issue has been resolved
by adding a unique constraint on codes. So, this
check is only for the existing data.