Add weekly/monthly automatic credits jobs to cron
These jobs were manually entered into the cron and
got overwritten. They should be in the schedule.rb,
so this doesn't happen again.

The monthly job was changed so that it runs at the
beginning of the month and runs the last month,
simply because it is much easier to setup a cron
entry to run on the first of the month, than dealing
with creating a cron job to run on the last day of the
month.